refactor(split): complete medium and low priority component splits
- BroadcastModal 833→55 lines: StreamTab+YouTubeTab extracted to broadcast/,
  StreamTab uses shared RelayPanel (removes duplicated RelayRow)
- SidebarLayout 709→95 lines: StatusPopover, QuickActionsPopover, TopBar,
  Sidebar/SidebarGroup/SidebarItem → sidebar/ with navConfig.js
- DskViewportsPage 862→493 lines: TextLayersEditor, ImageSettingsTable, ImageRow
  → dsk-viewports/ with shared styles.js
- CaptionsModal: VAD tab → panels/VadPanel.jsx
- Production pages: shared production/ConnectionDot.jsx
- Plan doc updated with actual line counts and status

Agent-Logs-Url: https://github.com/jsilvanus/live-captions-yt/sessions/13dcd3d8-2e20-4f71-851c-de23659db507
