+++ packages/lcyt-web/src/components/DskViewportsPage.jsx
@@ -1,64 +1,8 @@
 import { useContext, useEffect, useState, useCallback, useId } from 'react';
 import { SessionContext } from '../contexts/SessionContext';
-
-/**
- * DSK Viewports Management Page
- *
- * URL: /dsk-viewports?server=<backendUrl>&apikey=<key>
- *
- * Manage named viewport targets for DSK graphics:
- *   - Built-in "Landscape (default)" (1920×1080, not deletable — existing /dsk/:key behaviour)
- *   - User-defined viewports: vertical screens, viewer, etc.
- *
- * For each viewport:
- *   - Edit label, type, dimensions
- *   - Get the display URL (open in OBS / HDMI output browser)
- *   - "Present to screen" — open display URL on a specific connected monitor (Window Placement API)
- *   - Per-image settings: visible toggle, x/y/width/height position override, animation
- *
- * Auth: X-API-Key header (same as DskControlPage — no live session required).
- */
-
-// ── Shared styles ─────────────────────────────────────────────────────────────
-
-const dark = {
-  bg:       '#121212',
-  panel:    '#1a1a1a',
-  card:     '#222',
-  cardHover:'#2a2a2a',
-  border:   '#333',
-  text:     '#ddd',
-  muted:    '#888',
-  accent:   '#44ff88',
-  accentDim:'#2d8a52',
-  danger:   '#ff6666',
-};
-
-const btnBase = {
-  background: dark.card,
-  border: `1px solid ${dark.border}`,
-  color: dark.text,
-  borderRadius: 6,
-  padding: '7px 14px',
-  fontSize: 13,
-  cursor: 'pointer',
-  whiteSpace: 'nowrap',
-};
-const btnPrimary  = { ...btnBase, background: '#1a4a2e', border: `1px solid ${dark.accentDim}`, color: '#cfffdc' };
-const btnDanger   = { ...btnBase, background: '#3a0000', border: '1px solid #882222', color: '#ffaaaa' };
-const btnSmall    = { ...btnBase, padding: '4px 10px', fontSize: 12 };
-
-const inputStyle = {
-  background: '#1a1a1a',
-  border: `1px solid ${dark.border}`,
-  color: '#eee',
-  borderRadius: 4,
-  padding: '6px 10px',
-  fontSize: 13,
-  boxSizing: 'border-box',
-};
-
-const labelStyle = { display: 'block', marginBottom: 4, color: dark.muted, fontSize: 12 };
+import { dark, btnPrimary, btnDanger, btnSmall, inputStyle, labelStyle } from './dsk-viewports/styles.js';
+import { TextLayersEditor } from './dsk-viewports/TextLayersEditor.jsx';
+import { ImageSettingsTable } from './dsk-viewports/ImageSettingsTable.jsx';
 
 // ── Component ─────────────────────────────────────────────────────────────────
 
@@ -519,8 +463,8 @@ export function DskViewportsPage() {
   );
 }
 
-// Export ImageSettingsTable for reuse in the Editor
-export { ImageSettingsTable };
+// Re-export ImageSettingsTable from the dsk-viewports sub-module for reuse in the Editor
+export { ImageSettingsTable } from './dsk-viewports/ImageSettingsTable.jsx';
 
 // ── Sub-components ────────────────────────────────────────────────────────────
 
@@ -544,319 +488,6 @@ function UrlBlock({ url, onCopy }) {
       </code>
       <button style={btnSmall} onClick={copy}>Copy</button>
       <button style={btnSmall} onClick={() => window.open(url, '_blank')}>Open</button>
-    </div>
-  );
-}
-
-function ImageSettingsTable({ images, viewportName, getImgVpSettings, saveImgVpSettings, serverUrl }) {
-  const [expanded, setExpanded] = useState(new Set());
-
-  function toggle(id) {
-    setExpanded(s => {
-      const n = new Set(s);
-      n.has(id) ? n.delete(id) : n.add(id);
-      return n;
-    });
-  }
-
-  return (
-    <div>
-      {/* Header row */}
-      <div style={{ display: 'grid', gridTemplateColumns: '48px 1fr 80px 80px', gap: 8, padding: '4px 8px', fontSize: 11, color: dark.muted, borderBottom: `1px solid ${dark.border}`, marginBottom: 4 }}>
-        <span />
-        <span>Shorthand</span>
-        <span>Visible</span>
-        <span>Position</span>
-      </div>
-      {images.map(img => {
-        const settings = getImgVpSettings(img, viewportName);
-        const isExpanded = expanded.has(img.id);
-        return (
-          <ImageRow
-            key={img.id}
-            img={img}
-            settings={settings}
-            isExpanded={isExpanded}
-            onToggleExpand={() => toggle(img.id)}
-            onSave={patch => saveImgVpSettings(img, viewportName, patch)}
-            serverUrl={serverUrl}
-          />
-        );
-      })}
-    </div>
-  );
-}
-
-const COMMON_BINDINGS = ['section', 'stanza', 'speaker', 'song', 'lyrics'];
-
-function TextLayersEditor({ layers, onChange, vpWidth, vpHeight }) {
-  function updateLayer(idx, patch) {
-    onChange(ls => ls.map((l, i) => i === idx ? { ...l, ...patch } : l));
-  }
-  function removeLayer(idx) {
-    onChange(ls => ls.filter((_, i) => i !== idx));
-  }
-
-  if (layers.length === 0) {
-    return <div style={{ color: dark.muted, fontSize: 13 }}>No text layers. Click "+ Add Layer" to add one.</div>;
-  }
-
-  return (
-    <div>
-      {layers.map((layer, idx) => (
-        <div key={layer.id || idx} style={{ background: '#1a1a1a', border: `1px solid ${dark.border}`, borderRadius: 6, padding: 14, marginBottom: 10 }}>
-          <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', marginBottom: 10 }}>
-            <span style={{ fontSize: 13, fontWeight: 'bold', color: dark.text }}>
-              Layer {idx + 1}{layer.binding ? ` — bound to "${layer.binding}"` : ' — static text'}
-            </span>
-            <button style={{ ...btnDanger, padding: '3px 8px', fontSize: 11 }} onClick={() => removeLayer(idx)}>Remove</button>
-          </div>
-          <div style={{ display: 'flex', flexWrap: 'wrap', gap: 10 }}>
-            {/* Binding */}
-            <div style={{ flex: '1 1 200px' }}>
-              <label style={labelStyle}>Binding (code key)</label>
-              <div style={{ display: 'flex', gap: 6 }}>
-                <input
-                  list={`bindings-list-${idx}`}
-                  style={{ ...inputStyle, flex: 1 }}
-                  placeholder="section, stanza, speaker…"
-                  value={layer.binding || ''}
-                  onChange={e => updateLayer(idx, { binding: e.target.value })}
-                />
-                <datalist id={`bindings-list-${idx}`}>
-                  {COMMON_BINDINGS.map(b => <option key={b} value={b} />)}
-                </datalist>
-              </div>
-            </div>
-            {/* Static text (fallback / non-bound) */}
-            <div style={{ flex: '1 1 200px' }}>
-              <label style={labelStyle}>Static text (shown when no binding value)</label>
-              <input
-                style={{ ...inputStyle, width: '100%' }}
-                placeholder="optional fallback"
-                value={layer.text || ''}
-                onChange={e => updateLayer(idx, { text: e.target.value })}
-              />
-            </div>
-          </div>
-          <div style={{ display: 'flex', flexWrap: 'wrap', gap: 10, marginTop: 10 }}>
-            {[['x', 'X (px)'], ['y', 'Y (px)'], ['width', 'Width (px)'], ['height', 'Height (px)']].map(([key, lbl]) => (
-              <div key={key} style={{ flex: '0 0 100px' }}>
-                <label style={labelStyle}>{lbl}</label>
-                <input
-                  style={{ ...inputStyle, width: '100%' }}
-                  type="number"
-                  value={layer[key] ?? ''}
-                  onChange={e => updateLayer(idx, { [key]: parseInt(e.target.value, 10) || 0 })}
-                />
-              </div>
-            ))}
-            <div style={{ flex: '0 0 80px' }}>
-              <label style={labelStyle}>Font size</label>
-              <input
-                style={{ ...inputStyle, width: '100%' }}
-                type="number"
-                value={layer.fontSize ?? 48}
-                onChange={e => updateLayer(idx, { fontSize: parseInt(e.target.value, 10) || 48 })}
-              />
-            </div>
-            <div style={{ flex: '0 0 80px' }}>
-              <label style={labelStyle}>Font weight</label>
-              <select
-                style={{ ...inputStyle, width: '100%' }}
-                value={layer.fontWeight ?? 'bold'}
-                onChange={e => updateLayer(idx, { fontWeight: e.target.value })}
-              >
-                <option value="normal">Normal</option>
-                <option value="bold">Bold</option>
-                <option value="600">600</option>
-              </select>
-            </div>
-            <div style={{ flex: '0 0 80px' }}>
-              <label style={labelStyle}>Color</label>
-              <input
-                style={{ ...inputStyle, width: '100%', padding: '4px 6px' }}
-                type="color"
-                value={layer.color || '#ffffff'}
-                onChange={e => updateLayer(idx, { color: e.target.value })}
-              />
-            </div>
-            <div style={{ flex: '0 0 100px' }}>
-              <label style={labelStyle}>Align</label>
-              <select
-                style={{ ...inputStyle, width: '100%' }}
-                value={layer.textAlign ?? 'center'}
-                onChange={e => updateLayer(idx, { textAlign: e.target.value })}
-              >
-                <option value="left">Left</option>
-                <option value="center">Center</option>
-                <option value="right">Right</option>
-              </select>
-            </div>
-            <div style={{ flex: '1 1 180px' }}>
-              <label style={labelStyle}>Text shadow (CSS)</label>
-              <input
-                style={{ ...inputStyle, width: '100%' }}
-                placeholder="0 2px 8px rgba(0,0,0,0.8)"
-                value={layer.textShadow || ''}
-                onChange={e => updateLayer(idx, { textShadow: e.target.value })}
-              />
-            </div>
-          </div>
-          {/* Mini preview */}
-          {(vpWidth && vpHeight) && (
-            <div style={{ marginTop: 12 }}>
-              <TextLayerMiniPreview layer={layer} vpWidth={vpWidth} vpHeight={vpHeight} />
-            </div>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function TextLayerMiniPreview({ layer, vpWidth, vpHeight }) {
-  const PREVIEW_W = 280;
-  const scale = PREVIEW_W / vpWidth;
-  const previewH = vpHeight * scale;
-  return (
-    <div style={{ position: 'relative', width: PREVIEW_W, height: previewH, background: '#181818', border: `1px solid ${dark.border}`, borderRadius: 4, overflow: 'hidden' }}>
-      {/* Grid lines */}
-      <div style={{ position: 'absolute', inset: 0, backgroundImage: 'linear-gradient(rgba(255,255,255,0.04) 1px, transparent 1px), linear-gradient(90deg, rgba(255,255,255,0.04) 1px, transparent 1px)', backgroundSize: `${PREVIEW_W/8}px ${previewH/8}px` }} />
-      {/* Text layer preview box */}
-      <div style={{
-        position: 'absolute',
-        left:       (layer.x || 0) * scale,
-        top:        (layer.y || 0) * scale,
-        width:      (layer.width || 400) * scale,
-        height:     (layer.height || 80) * scale,
-        border:     '1px dashed rgba(68,255,136,0.6)',
-        background: 'rgba(68,255,136,0.05)',
-        display:    'flex',
-        alignItems: 'center',
-        justifyContent: layer.textAlign === 'right' ? 'flex-end' : layer.textAlign === 'center' ? 'center' : 'flex-start',
-        fontSize:   Math.max(8, (layer.fontSize || 48) * scale),
-        fontWeight: layer.fontWeight || 'bold',
-        color:      layer.color || '#ffffff',
-        padding:    '0 4px',
-        overflow:   'hidden',
-        whiteSpace: 'nowrap',
-      }}>
-        {layer.binding || layer.text || <span style={{ opacity: 0.4, fontStyle: 'italic' }}>binding</span>}
-      </div>
-    </div>
-  );
-}
-
-function ImageRow({ img, settings, isExpanded, onToggleExpand, onSave, serverUrl }) {
-  const visible   = settings.visible !== false;
-  const hasPos    = settings.x != null || settings.y != null || settings.width != null || settings.height != null;
-  const hasAnim   = !!settings.animation;
-
-  // Local draft for position editing
-  const [draft, setDraft] = useState({
-    x:         settings.x         != null ? String(settings.x)         : '',
-    y:         settings.y         != null ? String(settings.y)         : '',
-    width:     settings.width     != null ? String(settings.width)     : '',
-    height:    settings.height    != null ? String(settings.height)    : '',
-    animation: settings.animation ?? '',
-  });
-
-  // Sync draft when settings change (e.g. after save)
-  useEffect(() => {
-    setDraft({
-      x:         settings.x         != null ? String(settings.x)         : '',
-      y:         settings.y         != null ? String(settings.y)         : '',
-      width:     settings.width     != null ? String(settings.width)     : '',
-      height:    settings.height    != null ? String(settings.height)    : '',
-      animation: settings.animation ?? '',
-    });
-  }, [settings.x, settings.y, settings.width, settings.height, settings.animation]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  function parsePx(val) {
-    const n = parseInt(val, 10);
-    return isNaN(n) ? null : n;
-  }
-
-  function handleSave() {
-    onSave({
-      x:         parsePx(draft.x),
-      y:         parsePx(draft.y),
-      width:     parsePx(draft.width),
-      height:    parsePx(draft.height),
-      animation: draft.animation || null,
-    });
-  }
-
-  return (
-    <div style={{ borderBottom: `1px solid ${dark.border}` }}>
-      {/* Summary row */}
-      <div style={{ display: 'grid', gridTemplateColumns: '48px 1fr 80px 80px', gap: 8, padding: '6px 8px', alignItems: 'center' }}>
-        {/* Thumbnail */}
-        <img
-          src={`${serverUrl}/images/${img.id}`}
-          alt=""
-          style={{ width: 40, height: 40, objectFit: 'contain', background: '#111', borderRadius: 4, border: `1px solid ${dark.border}` }}
-          crossOrigin="anonymous"
-        />
-        {/* Shorthand */}
-        <div>
-          <code style={{ fontSize: 13, color: '#aef' }}>{img.shorthand}</code>
-          <span style={{ fontSize: 11, color: dark.muted, marginLeft: 8 }}>{img.mimeType?.split('/')[1]}</span>
-        </div>
-        {/* Visible toggle */}
-        <div>
-          <input
-            type="checkbox"
-            checked={visible}
-            onChange={e => onSave({ visible: e.target.checked })}
-            style={{ width: 16, height: 16, cursor: 'pointer' }}
-          />
-        </div>
-        {/* Expand button */}
-        <button
-          style={{ ...btnSmall, background: isExpanded ? dark.cardHover : 'transparent', border: `1px solid ${isExpanded ? dark.border : 'transparent'}` }}
-          onClick={onToggleExpand}
-          title="Position & animation overrides"
-        >
-          {hasPos || hasAnim ? '✎' : '+'} pos
-        </button>
-      </div>
-
-      {/* Expanded position row */}
-      {isExpanded && (
-        <div style={{ padding: '8px 56px 12px', background: '#1a1a1a', borderTop: `1px solid ${dark.border}` }}>
-          <div style={{ display: 'flex', flexWrap: 'wrap', gap: 10, marginBottom: 10 }}>
-            {[['x', 'X (px)'], ['y', 'Y (px)'], ['width', 'Width (px)'], ['height', 'Height (px)']].map(([key, lbl]) => (
-              <div key={key} style={{ flex: '0 0 100px' }}>
-                <label style={labelStyle}>{lbl}</label>
-                <input
-                  style={{ ...inputStyle, width: '100%' }}
-                  type="number"
-                  placeholder="auto"
-                  value={draft[key]}
-                  onChange={e => setDraft(d => ({ ...d, [key]: e.target.value }))}
-                />
-              </div>
-            ))}
-            <div style={{ flex: '1 1 200px' }}>
-              <label style={labelStyle}>Animation</label>
-              <input
-                style={{ ...inputStyle, width: '100%' }}
-                placeholder="e.g. lcyt-fadeIn 0.5s"
-                value={draft.animation}
-                onChange={e => setDraft(d => ({ ...d, animation: e.target.value }))}
-              />
-            </div>
-          </div>
-          <div style={{ display: 'flex', gap: 8 }}>
-            <button style={btnPrimary} onClick={handleSave}>Save</button>
-            <button style={btnDanger} onClick={() => onSave({ x: null, y: null, width: null, height: null, animation: null })}>
-              Reset position
-            </button>
-          </div>
-        </div>
-      )}
     </div>
   );
 }
