+++ packages/lcyt-web/src/components/ProductionBridgesPage.jsx
@@ -1,20 +1,7 @@
 import { KEYS } from '../lib/storageKeys.js';
 import { useState, useEffect, useCallback, useContext } from 'react';
 import { SessionContext } from '../contexts/SessionContext';
-
-function ConnectionDot({ connected }) {
-  return (
-    <span
-      title={connected ? 'Connected' : 'Disconnected'}
-      style={{
-        display: 'inline-block',
-        width: 8, height: 8, borderRadius: '50%', flexShrink: 0,
-        background: connected ? 'var(--color-success)' : 'var(--color-text-muted)',
-        boxShadow: connected ? '0 0 5px var(--color-success)' : 'none',
-      }}
-    />
-  );
-}
+import { ConnectionDot } from './production/ConnectionDot.jsx';
 
 function AddBridgeForm({ onCreated, onCancel, backendUrl, headers }) {
   const [name, setName] = useState('');
